+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial scaffolding of the provider based on v0.25.10 of the Snowflake Terraform Provider
 
 ---